+++ src/Products/PloneMeeting/tests/helpers.py
@@ -163,21 +163,22 @@ class PloneMeetingTestingHelpers:
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
         return getattr(self, ('TRANSITIONS_FOR_ACCEPTING_ITEMS_MEETING_%d' % meetingConfigNumber))
 
+    def get_transitions_for_proposing_item(self, first_level=False):
+        meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
+        if first_level:
+           return getattr(self,
+                   ('TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_%d' % meetingConfigNumber))
+        else:
+           return getattr(self,
+                   ('TRANSITIONS_FOR_PROPOSING_ITEM_%d' % meetingConfigNumber))
+
     def proposeItem(self, item, first_level=False):
         '''Propose passed p_item using TRANSITIONS_FOR_PROPOSING_ITEM_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_PROPOSING_ITEM_x is 1 or 2.
            If p_first_level is True, we will use TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_x,
            this makes it possible to reach an intermediate propose level.'''
-        meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        if first_level:
-            self._doTransitionsFor(item,
-                                   getattr(self,
-                                           ('TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_%d' % meetingConfigNumber)))
-        else:
-            self._doTransitionsFor(item,
-                                   getattr(self,
-                                           ('TRANSITIONS_FOR_PROPOSING_ITEM_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(item, self.get_transitions_for_proposing_item(first_level))
 
     def prevalidateItem(self, item):
         '''Prevalidate passed p_item using TRANSITIONS_FOR_PREVALIDATING_ITEM_x.
@@ -236,21 +237,26 @@ class PloneMeetingTestingHelpers:
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
         BACK_TO_WF_PATH = getattr(self, 'BACK_TO_WF_PATH_%d' % meetingConfigNumber)
         useDefinedWfPath = False
+        # check if a mapping exist for state name, returns original state if no mapping exists
+        state = self._stateMappingFor(state)
         if state in BACK_TO_WF_PATH:
             transitions = BACK_TO_WF_PATH[state]
             useDefinedWfPath = True
-        # check if a mapping exist for state name, returns original state if no mapping exists
-        state = self._stateMappingFor(state)
         # do things as admin to avoid permission issues
         currentUser = self.member.getId()
         self.changeUser('admin')
-        while not itemOrMeeting.queryState() == state:
+        max_attempts = 20
+        nb_attempts = 0
+        while not itemOrMeeting.queryState() == state or nb_attempts >= max_attempts:
+            nb_attempts += 1
             if not useDefinedWfPath:
                 transitions = self.transitions(itemOrMeeting)
             for tr in transitions:
                 if (tr.startswith('back') or useDefinedWfPath) and tr in self.transitions(itemOrMeeting):
                     self.do(itemOrMeeting, tr)
                     break
+        if nb_attempts >= max_attempts:
+            raise ValueError('impossible to go back to {}'.format(state))
         self.changeUser(currentUser)
 
     def _doTransitionsFor(self, itemOrMeeting, transitions):