+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -2800,6 +2800,64 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertRaises(Unauthorized, formAssembly.update)
         self.assertRaises(Unauthorized, formSignatures.update)
 
+    def test_pm_MayQuickEditItemAssemblyAndSignatures(self):
+        """Method that protects edition of itemAssembly/itemSignatures fields.
+           Only a MeetingManager may edit these fields until meeting is closed."""
+        def _checkOnlyEditableByManagers(item,
+                                         may_edit=['pmManager'],
+                                         may_not_edit=['pmCreator1', 'pmReviewer1']):
+            """ """
+            original_user_id = self.member.getId()
+            for user_id in may_edit:
+                self.changeUser(user_id)
+                self.assertTrue(item.mayQuickEditItemAssembly())
+                self.assertTrue(item.mayQuickEditItemSignatures())
+            for user_id in may_not_edit:
+                self.changeUser(user_id)
+                self.assertFalse(item.mayQuickEditItemAssembly())
+                self.assertFalse(item.mayQuickEditItemSignatures())
+            self.changeUser(original_user_id)
+
+        cfg = self.meetingConfig
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem')
+        item.setDecision(self.decisionText)
+        cfg.setUsedMeetingAttributes(('assembly', 'signatures'))
+        self.assertFalse(item.mayQuickEditItemAssembly())
+        self.assertFalse(item.mayQuickEditItemSignatures())
+        self.validateItem(item)
+        self.changeUser('pmReviewer1')
+        self.assertFalse(item.mayQuickEditItemAssembly())
+        self.assertFalse(item.mayQuickEditItemSignatures())
+        # only editable when in meeting
+        self.changeUser('pmManager')
+        self.assertFalse(item.mayQuickEditItemAssembly())
+        self.assertFalse(item.mayQuickEditItemSignatures())
+        meeting = self.create('Meeting', date=DateTime('2020/02/10'))
+        self.presentItem(item)
+        _checkOnlyEditableByManagers(item)
+        # decide meeting
+        self.decideMeeting(meeting)
+        _checkOnlyEditableByManagers(item)
+        # accept item
+        self.do(item, 'accept')
+        _checkOnlyEditableByManagers(item)
+        # if not used, fields are not editable
+        cfg.setUsedMeetingAttributes(())
+        _checkOnlyEditableByManagers(item,
+                                     may_edit=[],
+                                     may_not_edit=['pmManager', 'pmCreator1', 'pmReviewer1'])
+        cfg.setUsedMeetingAttributes(('assembly', 'signatures'))
+        # change itemAssembly/itemSignatures
+        item.setItemAssembly('New assembly')
+        item.setItemSignatures('New signatures')
+        _checkOnlyEditableByManagers(item)
+        # no more editable by anybody when meeting closed
+        self.closeMeeting(meeting)
+        _checkOnlyEditableByManagers(item,
+                                     may_edit=[],
+                                     may_not_edit=['pmManager', 'pmCreator1', 'pmReviewer1'])
+
     def test_pm_Validate_item_assembly(self):
         """Test the method that validates item_assembly on the item_assembly_form.
            The validator logic is tested in testUtils.test_pm_Validate_item_assembly_value,
@@ -5197,6 +5255,30 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertEqual(helper.printAssembly(striked=True),
                          '<p>Assembly with <strike>striked</strike> part</p>')
 
+    def test_pm_printAssembly(self):
+        # Set up
+        self.changeUser('siteadmin')
+        cfg = self.meetingConfig
+        cfg.setUsedMeetingAttributes(('attendees', 'excused', 'absents', 'signatories',))
+        ordered_contacts = cfg.getField('orderedContacts').Vocabulary(cfg).keys()
+        cfg.setOrderedContacts(ordered_contacts)
+        self.changeUser('pmManager')
+        meeting = self.create('Meeting', date=DateTime())
+        item = self.create('MeetingItem')
+        self.presentItem(item)
+        template = self.meetingConfig.podtemplates.itemTemplate
+        self.request.set('template_uid', template.UID())
+        self.request.set('output_format', 'odt')
+        view = item.restrictedTraverse('@@document-generation')
+        view()
+        helper = view.get_generation_context_helper()
+
+        printed_assembly = helper.printAssembly()
+        # Every attendee firstname and lastname must be in view.printAssembly()
+        for attendee in item.getAttendees(theObjects=True):
+            self.assertIn(attendee.get_person().firstname, printed_assembly)
+            self.assertIn(attendee.get_person().lastname, printed_assembly)
+
     def test_pm_DownOrUpWorkflowAgain(self):
         """Test the MeetingItem.downOrUpWorkflowAgain behavior."""
         self.changeUser('siteadmin')
@@ -5463,7 +5545,7 @@ class testMeetingItem(PloneMeetingTestCase):
         cfg = self.meetingConfig
         self.changeUser('pmCreator1')
         # creation time
-        text = '<p>Working external image <img src="http://www.imio.be/contact.png"/>.</p>'
+        text = '<p>Working external image <img src="https://i.picsum.photos/id/22/400/400.jpg"/>.</p>'
         pmFolder = self.getMeetingFolder()
         # do not use self.create to be sure that it works correctly with invokeFactory
         itemId = pmFolder.invokeFactory(cfg.getItemTypeName(),
@@ -5473,29 +5555,29 @@ class testMeetingItem(PloneMeetingTestCase):
         item = getattr(pmFolder, itemId)
         item.processForm()
         # contact.png was saved in the item
-        self.assertTrue('contact.png' in item.objectIds())
-        img = item.get('contact.png')
+        self.assertTrue('22-400x400.jpg' in item.objectIds())
+        img = item.get('22-400x400.jpg')
         # external image link was updated
         self.assertEqual(
             item.getRawDescription(),
             '<p>Working external image <img src="resolveuid/{0}">.</p>'.format(img.UID()))
 
         # test using the quickedit, test with field 'decision' where getRaw was overrided
-        decision = '<p>Working external image <img src="http://www.imio.be/mascotte-presentation.jpg"/>.</p>'
+        decision = '<p>Working external image <img src="https://i.picsum.photos/id/1025/400/300.jpg"/>.</p>'
         setFieldFromAjax(item, 'decision', decision)
-        self.assertTrue('mascotte-presentation.jpg' in item.objectIds())
-        img2 = item.get('mascotte-presentation.jpg')
+        self.assertTrue('1025-400x300.jpg' in item.objectIds())
+        img2 = item.get('1025-400x300.jpg')
         # external image link was updated
         self.assertEqual(
             item.getRawDecision(),
             '<p>Working external image <img src="resolveuid/{0}">.</p>'.format(img2.UID()))
 
         # test using at_post_edit_script, aka full edit form
-        decision = '<p>Working external image <img src="http://www.imio.be/spw.png"/>.</p>'
+        decision = '<p>Working external image <img src="https://i.picsum.photos/id/1035/600/400.jpg"/>.</p>'
         item.setDecision(decision)
         item._update_after_edit()
-        self.assertTrue('spw.png' in item.objectIds())
-        img3 = item.get('spw.png')
+        self.assertTrue('1035-600x400.jpg' in item.objectIds())
+        img3 = item.get('1035-600x400.jpg')
         # external image link was updated
         self.assertEqual(
             item.getRawDecision(),
@@ -5503,14 +5585,14 @@ class testMeetingItem(PloneMeetingTestCase):
 
         # link to unknown external image, like during copy/paste of content
         # that has a link to an unexisting image or so
-        decision = '<p>Not working external image <img src="http://www.imio.be/unknown_image.png">.</p>'
+        decision = '<p>Not working external image <img src="https://i.picsum.photos/id/449/400.png">.</p>'
         item.setDecision(decision)
         item._update_after_edit()
-        self.assertTrue('spw.png' in item.objectIds())
+        self.assertTrue('1035-600x400.jpg' in item.objectIds())
         # nothing was done
-        self.assertEqual(
+        self.assertListEqual(
             sorted(item.objectIds()),
-            ['contact.png', 'mascotte-presentation.jpg', 'spw.png'])
+            ['1025-400x300.jpg', '1035-600x400.jpg', '22-400x400.jpg'])
         self.assertEqual(item.getRawDecision(), decision)
 
     def test_pm_ItemInternalImagesStoredLocallyWhenItemDuplicated(self):
@@ -5534,19 +5616,19 @@ class testMeetingItem(PloneMeetingTestCase):
             '<p>Internal image <img src="{1}">.</p>' \
             '<p>Internal image 2 <img src="{2}">.</p>'
         text = text_pattern.format(
-            'http://www.imio.be/contact.png',
+            'https://i.picsum.photos/id/1025/400/300.jpg', # 1025-400x300.jpg
             img.absolute_url(),
             'resolveuid/{0}'.format(img2.UID()))
         item.setDescription(text)
         self.assertEqual(item.objectIds(), ['dot.gif', 'dot2.gif'])
         item._update_after_edit()
         # we have images saved locally
-        self.assertEqual(sorted(item.objectIds()), ['contact.png', 'dot.gif', 'dot2.gif'])
+        self.assertEqual(sorted(item.objectIds()), ['1025-400x300.jpg', 'dot.gif', 'dot2.gif'])
 
         # duplicate and check that uri are correct
         newItem = item.clone()
-        self.assertEqual(sorted(newItem.objectIds()), ['contact.png', 'dot.gif', 'dot2.gif'])
-        new_img = newItem.get('contact.png')
+        self.assertEqual(sorted(newItem.objectIds()), ['1025-400x300.jpg', 'dot.gif', 'dot2.gif'])
+        new_img = newItem.get('1025-400x300.jpg')
         new_img1 = newItem.get('dot.gif')
         new_img2 = newItem.get('dot2.gif')
         # every links are turned to resolveuid